+++ src/utils.ts
@@ -1,5 +1,8 @@
-import path from 'path';
+import * as path from 'path';
 import * as vscode from 'vscode';
+import { TFolder } from '.';
+import crypto from 'crypto';
+import fileSystem from 'fs';
 
 export function getCwd() {
 	if (!vscode.workspace.workspaceFolders) {
@@ -11,15 +14,14 @@ export function getCwd() {
 	return cwd;
 }
 
-export function getNonce() {
+function getNonce() {
 	let text = '';
 	const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-	for (let i = 0; i < 32; i++) {
+	for (let i = 0; i < 16; i++) {
 		text += possible.charAt(Math.floor(Math.random() * possible.length));
 	}
 	return text;
 }
-
 export const nonce = getNonce();
 
 export function getGlob() {
@@ -68,39 +70,67 @@ export function getGlob() {
 }
 
 export function getFilename(imgPath: string) {
-	let filename = decodeURI(imgPath).split("/").pop();
-
+	const filename = decodeURI(imgPath).split("/").pop();
 	if (filename) {
 		return filename.split("?").shift();
 	}
 	return filename;
 }
 
-export type TypeOfImagesInSubFolders = {
-	imgUri: vscode.Uri,
-	imgMetadata: vscode.FileStat | null,
-};
+export function hash256(str: string, truncate = 16) {
+	return 'H' + crypto.createHash('sha256').update(str).digest('hex').substring(0, truncate);
+}
 
-export type TypeOfImagesBySubFolders = {
-	[key: string]: Array<TypeOfImagesInSubFolders>,
-};
+export async function getFileStats(imgUris: vscode.Uri[]) {
+	const result = await Promise.all(imgUris.map(async (imgUri: { fsPath: any; }) => {
+		var path = imgUri.fsPath;
+		var stat = await fileSystem.promises.stat(path);
+		return [path, stat];
+	}));
 
-export async function getImagesBySubFolders(imgPaths: vscode.Uri[], action: string = 'create') {
-	let imagesBySubFolders: TypeOfImagesBySubFolders = {};
-	let imgMetadata = null;
-	for (const imgUri of imgPaths) {
-		let key = path.dirname(imgUri.fsPath);
-		key = key[0].toUpperCase() + key.slice(1,);
-		if (!imagesBySubFolders[key]) {
-			imagesBySubFolders[key] = [];
-		}
-		if (action !== 'delete') {
-			imgMetadata = await vscode.workspace.fs.stat(imgUri);
-		}
-		imagesBySubFolders[key].push({
-			"imgUri": imgUri,
-			"imgMetadata": imgMetadata
-		});
+	const resultObj = result.reduce((obj, item) => {
+		return {
+			...obj,
+			[item[0]]: item[1],
+		};
+	}, {});
+
+	return resultObj;
+}
+
+export async function getFolders(imgUris: vscode.Uri[], action: "create" | "change" | "delete" = "create") {
+	let folders: Record<string, TFolder> = {};
+
+	let fileStats;
+	if (action !== "delete") {
+		fileStats = await getFileStats(imgUris);
 	}
-	return imagesBySubFolders;
+	for (const imgUri of imgUris) {
+		const folderPath = path.dirname(imgUri.path);
+		const folderId = hash256(folderPath);
+
+		if (!folders[folderId]) { // first image of the folder
+			folders[folderId] = {
+				id: folderId,
+				path: folderPath,
+				images: {},
+			};
+		}
+
+		if (action !== 'delete' && fileStats !== undefined) {
+			const fileStat = fileStats[imgUri.fsPath as keyof typeof fileStats];
+			const dotIndex = imgUri.fsPath.lastIndexOf('.');
+			const imageId = hash256(imgUri.path);
+			folders[folderId].images[imageId] = {
+				id: imageId,
+				uri: imgUri,
+				ext: imgUri.fsPath.slice(dotIndex + 1).toUpperCase(),
+				size: fileStat['size'],
+				mtime: fileStat['mtime'],
+				ctime: fileStat['ctime'],
+				status: "",
+			};
+		}
+	}
+	return folders;
 }